CMake: export cmake-configuration for the library
Add support to export the CMake-configuration during install.  This generates a
file `libcaerConfig.cmake` which provides `imported targets` for the library and
a `CMake`-variable for the include directories.  The config-file can be used by
client code to find the installed `libcaer` package using

    find_package(libcaer::caer REQUIRED)
    include_directories(${libcaer_INCLUDE_DIRS)

in the `CMakeLists.txt` without needing to set the include- and link-path by
hand.